requirements.txt
streamlit
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
pypdf
huggingface_hub
requests